Added data normalization for blank cookie fillings
This bugfix corrects an issue that was causing cookies without fillings
to incorrectly display no text at all instead of the desired
"no fillings" in the list of available cookies for a store.

The core of the issue was that the cookie creation form was passing an
empty string to the controller, but the view was handling for a
boolean when determining whether or not a cookie had fillings.

In order to correct this, a before_validation hook was added to the
cookie model to ensure that all fillings that match ruby's "blank?"
method will be converted to nil.  The problem could have been fixed
within the view, but this hopefully will avoid a resurgence of the bug
in other locations that might display cookie fillings similarly.