Add async polling for live updates upon baking completion
render_async gem added to allow more seamless async polling of the db
without needing to write custom js/ajax. The gem allows the controller
to render a partial after a call to the baking_status route.

The implementation itself is relatively straight forward. render_async
handles the rendering of the cookie-info partial, and checks for updates
from the database every 5 seconds. If there's a change, it will
re-render the partial, which in the case of this feature means that
it will update the oven show page with the retrieve cookie button.

For self-testing, I would recommend changing the OvenWorker's call to
Kernel.sleep from 120 seconds to something shorter like 5 or 10 seconds,
that way you can see the live update several times without having to
wait too long.