Add background worker to bake cookies async
Added sidekiq and redis as dependencies to enable asynchronous baking
of cookies.  Each created cookie now starts a background job through
OvenWorker that starts a 2 minute waiting period via Kernel.sleep.

After the 2 minutes, the cookie is flagged as baked via an associated
DateTime that is associated with the cookie's completed_baking_at
attribute.

The cookie.ready? instance method was also modified so that it
references this new field to check for a value instead of just returning
true in all cases.

Environment configurations for the background workers are not
implemented in this feature.  Redis will need to be running locally to
the rails server and sidekiq with the default port of 6379.