Add capability to bake multiple cookies in a single batch
This feature allows a user to create multiple cookies concurrently in
a single oven.  For simplicity, the user can opt to create 0 cookies in
the UI, and they will be returned to the oven show page with no new
cookies in the oven.  The user will have to purposefully select 0,
however, as the number of cookies per batch is a required field.

A baking completion field had to be added to ovens to ensure the full
batch could be properly tracked, as previously the completion of baking
was tracked by the single cookie being baked.

Some existing test setups had to be modified to pass the adjustments
made in this feature, since old behavior was altered.